+++ motulator/control/im/__init__.py
@@ -5,13 +5,18 @@ from motulator.control.im._vector import (
     CurrentReferencePars,
     CurrentReference,
     ModelPars,
-    Observer,
     VectorCtrl,
 )
 
+from motulator.control.im._observers import (Observer, FluxObserver)
+
 from motulator.control.im._obs_vhz import (
-    FluxObserver, ObserverBasedVHzCtrl, ObserverBasedVHzCtrlPars)
+    ObserverBasedVHzCtrl, ObserverBasedVHzCtrlPars)
 
 from motulator.control.im._vhz import VHzCtrl
 
-# from motulator.control._common import RateLimiter, SpeedCtrl
+__all__ = [
+    "CurrentCtrl", "CurrentReferencePars", "CurrentReference", "ModelPars",
+    "VectorCtrl", "Observer", "FluxObserver", "ObserverBasedVHzCtrl",
+    "ObserverBasedVHzCtrlPars", "VHzCtrl"
+]
