+++ motulator/control/im/_vector.py
@@ -8,6 +8,7 @@ from dataclasses import dataclass, InitVar
 import numpy as np
 from motulator._helpers import abc2complex
 from motulator.control._common import PWM, Ctrl, ComplexPICtrl, SpeedCtrl
+from motulator.control.im._observers import Observer, FullOrderObserver
 from motulator._utils import Bunch
 
 
@@ -349,145 +350,3 @@ class CurrentReference:
         u_s_max = self.k_u*u_dc/np.sqrt(3)
         self.i_sd_ref += T_s*self.k_fw*(u_s_max - np.abs(u_s_ref))
         self.i_sd_ref = np.clip(self.i_sd_ref, -self.i_s_max, self.i_sd_nom)
-
-
-# %%
-class Observer:
-    """
-    Reduced-order flux observer for induction machines.
-
-    This class implements a reduced-order flux observer for induction machines.
-    Both sensored and sensorless operation are supported. The observer structure
-    is similar to [#Hin2010]_. Both sensored and sensorless operation are 
-    supported. The observer operates in estimated rotor flux coordinates. 
-    
-    Parameters
-    ----------
-    par : ModelPars
-        Machine model parameters.
-    k : callable, optional
-        Observer gain as a function of the rotor angular speed. The default is 
-        ``lambda w_m: (0.5*R_R/L_M + 0.2*abs(w_m))/(R_R/L_M - 1j*w_m)`` if
-        `sensorless` else ``lambda w_m: 1 + 0.2*abs(w_m)/(R_R/L_M - 1j*w_m)``.
-    alpha_o : float, optional
-        Observer bandwidth (rad/s). The default is 2*pi*40.
-    sensorless : bool, optional
-        If True, sensorless mode is used. The default is True.
-
-    Attributes
-    ----------
-    psi_R : float
-        Rotor flux magnitude estimate (Vs).
-    theta_s : float
-        Rotor flux angle estimate (rad).
-    w_m : float
-        Rotor angular speed estimate (electrical rad/s). In sensored mode, this
-        is the measured low-pass-filtered speed.
-
-    Notes
-    -----
-    The pure voltage model corresponds to ``k = lambda w_m: 0``, resulting in 
-    the marginally stable estimation-error dynamics. The current model is 
-    obtained by setting ``k = lambda w_m: 1``. 
-
-    References
-    ----------
-    .. [#Hin2010] Hinkkanen, Harnefors, Luomi, "Reduced-order flux observers 
-       with stator-resistance adaptation for speed-sensorless induction motor 
-       drives," IEEE Trans. Power Electron., 2010, 
-       https://doi.org/10.1109/TPEL.2009.2039650
-
-    """
-
-    # pylint: disable=too-many-instance-attributes
-    def __init__(self, par, k=None, alpha_o=2*np.pi*40, sensorless=True):
-        # Model parameters
-        self.R_s, self.R_R, self.L_sgm = par.R_s, par.R_R, par.L_sgm
-        self.alpha = par.R_R/par.L_M
-
-        # Other parameters
-        self.alpha_o = alpha_o
-        self.sensorless = sensorless
-
-        # Observer gains
-        self.k1 = k
-        if self.sensorless:
-            if self.k1 is None:  # If not given, use the default gains
-                self.k1 = lambda w_m: (.5*self.alpha + .2*np.abs(w_m))/(
-                    self.alpha - 1j*w_m)
-            self.k2 = self.k1
-        else:
-            if self.k1 is None:
-                self.k1 = lambda w_m: 1 + .2*np.abs(w_m)/(self.alpha - 1j*w_m)
-            self.k2 = 0*self.k1
-
-        # States
-        self.psi_R, self.theta_s, self.w_m, self._i_s_old = 0, 0, 0, 0
-
-    def _f(self, T_s, u_s, i_s, w_m):
-        # Right-hand-side of the differential equation.
-
-        # Induced voltage from stator quantities (without the w_s term that is
-        # taken into account separately to avoid an algebraic loop)
-        v_s = u_s - self.R_s*i_s - self.L_sgm*(i_s - self._i_s_old)/T_s
-
-        # Induced voltage from the rotor quantities
-        v_r = self.R_R*i_s - (self.alpha - 1j*w_m)*self.psi_R
-
-        # Angular frequency of the rotor flux vector
-        k1, k2 = self.k1(w_m), self.k2(w_m)
-        den = self.psi_R + self.L_sgm*np.real((1 - k1)*i_s + k2*np.conj(i_s))
-        num = np.imag(v_s + k1*(v_r - v_s) + k2*np.conj(v_r - v_s))
-        w_s = num/den if den > 0 else w_m
-
-        # Compute the derivative of the flux magnitude for the state update
-        v = v_s - 1j*w_s*self.L_sgm*i_s
-        dpsi_R = np.real(v + k1*(v_r - v) + k2*np.conj(v_r - v))
-
-        return dpsi_R, w_s  # Note: w_s = dtheta_s
-
-    # pylint: disable=too-many-arguments
-    def _update(self, T_s, i_s, dpsi_R, w_s, w_m):
-        # Update the states
-        self.psi_R += T_s*dpsi_R
-        self.theta_s += T_s*w_s  # Next line: limit into [-pi, pi)
-        self.theta_s = np.mod(self.theta_s + np.pi, 2*np.pi) - np.pi
-        self.w_m += T_s*self.alpha_o*(w_m - self.w_m)
-        # Store the old current
-        self._i_s_old = i_s
-
-    def __call__(self, T_s, u_s, i_s, w_m=None):
-        """
-        Compute the angular frequency of the flux and update the states.
-
-        Parameters
-        ----------
-        T_s : float
-            Sampling period (s).
-        u_s : complex
-            Stator voltage (V) in estimated rotor flux coordinates.
-        i_s : complex
-            Stator current (A) in estimated rotor flux coordinates.
-        w_m : float, optional
-            Rotor angular speed (electrical rad/s). This signal is only used in
-            sensored mode. The default is None.
-
-        Returns
-        -------
-        w_s : float
-            Angular frequency (rad/s) of the rotor flux estimate.
-
-        """
-        # Compute the time derivative of the flux estimate
-        if self.sensorless:
-            dpsi_R, w_s = self._f(T_s, u_s, i_s, self.w_m)
-            # Slip and unfiltered speed estimates
-            w_r = self.R_R*i_s.imag/self.psi_R if self.psi_R > 0 else 0
-            w_m = w_s - w_r
-        else:
-            dpsi_R, w_s = self._f(T_s, u_s, i_s, w_m)
-
-        # Update the states
-        self._update(T_s, i_s, dpsi_R, w_s, w_m)
-
-        return w_s
